Rename mock.ts to mockClient.ts
