+++ shared/spotify-api/index.ts
@@ -1,4 +1,4 @@
 export { Spotify } from './src/types';
 export { OAUTH_SCOPES } from './src/constants';
 export { SpotifyDataSource } from './src/dataSource';
-export { MockSpotifyClient } from './src/mock';
+export { MockSpotifyClient } from './src/mockClient';
