+++ shared/spotify-api/src/mockClient.ts
